% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxPF-package.R
\docType{package}
\name{CoxPF}
\alias{CoxPF}
\alias{CoxPF-package}
\title{CoxPF: Profile Monitoring Methods for Cox Modeling}
\description{
Functions to implement nonparametric profile monitoring methods for the
cumulative baseline hazard for Cox models.
Major quantities that are computed:
\itemize{
\item reference shape when a matrix of shapes are given
\item reference variabilities after the reference shape is subtracted
from all shapes
}
}
