% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genDataWeibull}
\alias{genDataWeibull}
\title{Generate Survival Data}
\usage{
genDataWeibull(weibA, weibB, beta, n, tmax, epsilon)
}
\arguments{
\item{weibA}{Weibull shape parameter}

\item{weibB}{Weibull rate parameter}

\item{beta}{vector of coefficients}

\item{n}{number of observations}

\item{tmax}{maximum follow-up time}

\item{epsilon}{controls the censoring rate}
}
\value{
a dataframe of five columns:
\itemize{
\item \code{survtime} - survival times
\item \code{status} - final statuses
\item \code{age} - center and scaled age
\item \code{Sex} - gender indicator
\item \code{Black} - race indicator
}
}
\description{
Generate Survival Data
}
