% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPhaseI.R
\name{genPhaseI}
\alias{genPhaseI}
\title{Generate Phase I Profiles and Their Summaries}
\usage{
genPhaseI(nprofile, baseline, alphalevel)
}
\arguments{
\item{nprofile}{Number of Phase I profiles}

\item{baseline}{\code{"exp"} for exponential baseline hazard, \code{"weib"} for Weibull
baseline hazard}

\item{alphalevel}{overall alpha level for Phase I in determining the thresholds,
corresponding to \eqn{\alpha_0} in Equation 13 in the manuscript.}
}
\value{
a list of items:
\itemize{
\item alpha - the respective alpha levels chosen for each of the four
statistics
\item d.upper - the control limit for vertical shift
\item T1.upper - the control limit for max local shape deviation
\item T2.upper - the control limit for cumulative shape deviation
\item t2.upper - the control limit for Hotelling's \eqn{T^2}; if \code{nprofile < 100},
it is the empirical quantile for Phase I \eqn{T^2} statistics; otherwise it
is the quantile of \eqn{\chi^2_3} distribution corresponding to alpha level
\item betaK - the weighted average beta vector from Phase I
\item muhatdelta - the median of Phase I profile centers
\item shatdelta - the median absolute deviation of Phase I profile centers
\item refShape - the estimated reference shape curve
\item refVariability - the estimated reference variability curve
\item refCor - the reference correlation matrix of Phase I covariates
}
}
\description{
Generate Phase I Profiles and Their Summaries
}
