% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfile.R
\name{getProfile}
\alias{getProfile}
\title{Obtain the Cumulative Baseline Hazard for A Given Time Grid}
\usage{
getProfile(coxFit, times, newdata)
}
\arguments{
\item{coxFit}{A \code{coxph} object}

\item{times}{The vector of time points where the cumulative baseline hazards
are to be estimated}

\item{newdata}{A new dataset containing the observation whose cumulative
baseline hazard we want to predict. If this argument is not given, an
hypothetical observation whose covariates equal the average of covariates
for all observations in the original data will be used.}
}
\value{
A list of four objects:
\itemize{
\item \code{bh.orig} -  the estimated cumulative baseline hazard
\item \code{bh.std} - the pointwise standard deviation of \code{bh.orig}
\item \code{beta.orig} - the coefficient vector from the Cox regression
\item \code{beta.var} - the variance-covariance matrix of estimated coefficients
}
}
\description{
This function takes a \code{coxph} object, and extracts the cumulative baseline
hazard.
}
