% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lad_variability}
\alias{lad_variability}
\title{Compute a reference variability curve for a collection of profiles}
\usage{
lad_variability(cprofileMat, bandwidth, gridprecision, timevec)
}
\arguments{
\item{cprofileMat}{a matrix of profiles with zero medians}

\item{bandwidth}{the bandwidth used in RBF kernel}

\item{gridprecision}{the precision for grid search}

\item{timevec}{vector of time}
}
\value{
a vector whose length equals the length of time vector containing
the reference variability at each time point
}
\description{
Compute a reference variability curve for a collection of profiles
}
