% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lads.R
\name{lads}
\alias{lads}
\title{Obtain the LAD Estimations of Reference Shape and Variability}
\usage{
lads(profileMat, bw1, bw2, pc1, pc2, timevec)
}
\arguments{
\item{profileMat}{A matrix or data.frame, each line of which contain one
set of measurements on the time grid}

\item{bw1}{Bandwidth for RBF kernel used in estimating the reference
shape}

\item{bw2}{Bandwidth for RBF kernel used in estimating the reference
variability}

\item{pc1}{Precision of the grid to use in estimating the reference
shape}

\item{pc2}{Precision of the grid to use in estimating the reference
variability}

\item{timevec}{The vector of time points where the reference shape and
variability curves are to be estimated}
}
\value{
A list of three objects
\itemize{
\item \code{locations} - the medians for each profile
\item \code{shapes} - the reference shape
\item \code{var} - the reference variability
}
}
\description{
This is a wrapper funciton to obtain the least absolute deviation estimates
of reference shape and variability when we have a number of measurements,
each of which consist of (time/location, value) pairs.
}
