% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladStats.R
\name{ladStats}
\alias{ladStats}
\title{Compute Three Nonparametric Descriptive Statistics}
\usage{
ladStats(profileMat, bw1, bw2, pc1, pc2, timevec)
}
\arguments{
\item{profileMat}{A matrix or data.frame, each line of which contain one
set of measurements on the time grid, after adjusting for pointwise
variation}

\item{bw1}{Bandwidth for RBF kernel used in estimating the reference
shape}

\item{bw2}{Bandwidth for RBF kernel used in estimating the reference
variability}

\item{pc1}{Precision of the grid to use in estimating the reference
shape}

\item{pc2}{Precision of the grid to use in estimating the reference
variability}

\item{timevec}{The vector of time points where the reference shape and
variability curves are to be estimated}
}
\value{
A list:
\itemize{
\item \code{D}  - profile centers
\item \code{T1} - local shape deviation
\item \code{T2} - cumulative shape deviation
\item \code{shape} - the reference shape
\item \code{variability} - the reference variability
}
}
\description{
Given a matrix of profiles, this function fits the LAD analysis, computes
the reference shape and variation curves, and returns the profile centers
and the matrix of relative shift. It works on top of \code{\link{lads}}.
}
\seealso{
\code{\link{lads}}
}
