% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpf-package.R
\docType{package}
\name{coxpf}
\alias{coxpf}
\alias{coxpf-package}
\title{coxpf: Profile Monitoring Methods for Cox Modeling}
\description{
Functions to implement nonparametric profile monitoring methods for the
cumulative baseline hazard for Cox models.
Major quantities that are computed:
\itemize{
\item reference shape when a matrix of shapes are given
\item reference variabilities after the reference shape is subtracted
from all shapes
}
}
